+++ plone/app/event/__init__.py
@@ -2,8 +2,7 @@ from Products.CMFCore.permissions import setDefaultRoles
 from zope.i18nmessageid import MessageFactory
 
 packageName = __name__
-messageFactory = MessageFactory(packageName)
-PloneMessageFactory = MessageFactory('plone')
+_ = MessageFactory('plone')
 
 # BBB Permissions
 PORTAL_ADD_PERMISSION = 'Add portal events'  # CMFCalendar/ATCT permissions