+++ plone/app/event/vocabularies.py
@@ -1,5 +1,5 @@
 from plone.app.event import base
-from plone.app.event import messageFactory as _
+from plone.app.event import _
 from zope.interface import provider
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm